cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "devboardfinder"
version = "0.1.0"
description = "CLI for the Universal Board Description Standard (UBDS)"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "rich>=13.0",
    "requests>=2.28",
]

[project.scripts]
dbf = "dbf.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-tmp-files>=0.0.2"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
